extract preferences form wrapper
