+++ src/pages/preferences.tsx
@@ -1,63 +1,13 @@
-import {
-    Box,
-    Heading,
-    VStack,
-    useColorModeValue,
-    Progress,
-    Text,
-} from "@chakra-ui/react";
-import PreferencesForm from "../components/Preferences/PreferencesForm";
-import {
-    Difficulty,
-    Willingness,
-} from "../components/Preferences/CoursesPreferencesTable";
-import { useGetQuery } from "@hooks/useRequest";
+import dynamic from "next/dynamic";
 
-const getCourses = () => {
-    return ["CSC 225", "CSC 226", "ECE 260", "ECE 310", "SENG 265", "SENG 310"];
-};
+import { Box, Heading, VStack, useColorModeValue } from "@chakra-ui/react";
 
-const coursePreferencesInit = getCourses().reduce((obj, course) => {
-    obj[course] = {
-        willingness: Willingness.noSelection,
-        difficulty: Difficulty.noSelection,
-    };
-    return obj;
-}, {});
-
-const defaultInitialValues = {
-    nonTeachingSemester: "fall",
-    numCoursesPerSem: {
-        fall: 0,
-        spring: 0,
-        summer: 0,
-    },
-    sabbatical: {
-        value: false,
-        duration: "half",
-        fromMonth: "january",
-    },
-    teachingDaysPerWeek: {
-        fall: 0,
-        spring: 0,
-        summer: 0,
-    },
-    preferredDays: [],
-    preferredTime: {
-        fall: [],
-        summer: [],
-        spring: [],
-    },
-    coursePreferences: coursePreferencesInit,
-};
+const DynamicPreferencesFormWrapper = dynamic(
+    () => import("@components/Preferences/PreferencesFormWrapper"),
+    { ssr: false }
+);
 
 const Preferences = () => {
-    const { data, isError, isLoading } = useGetQuery("/api/preferences/");
-
-    console.log({ data, isError, isLoading });
-
-    const initialValues = data || defaultInitialValues;
-
     return (
         <VStack width="50%">
             <Heading mr="auto" mb={5}>
@@ -71,16 +21,7 @@ const Preferences = () => {
                 borderRadius={10}
                 boxShadow="rgba(99, 99, 99, 0.2) 0px 2px 8px 0px;"
             >
-                {isLoading && <Progress isIndeterminate />}
-                {isError && (
-                    <Text fontSize="sm" color="red.500">
-                        There was an error loading your preferences.
-                    </Text>
-                )}
-                <PreferencesForm
-                    isDisabled={isLoading}
-                    initialValues={initialValues}
-                />
+                <DynamicPreferencesFormWrapper />
             </Box>
         </VStack>
     );
